pyproject.toml
[tool.poetry]
name = "todowoo-drf"
version = "0.1.0"
description = ""
authors = ["Khasanjon Mamadaliyev <khasanjon.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
